add styles folder
