+++ frontend/src/pages/Login.tsx
@@ -1,7 +1,6 @@
 import 'react-native-gesture-handler';
 import React, {useState} from 'react';
 import {
-  StyleSheet,
   View,
   Text,
   TextInput,
@@ -9,6 +8,7 @@ import {
   Alert,
 } from 'react-native';
 import {Checkbox, Provider as PaperProvider} from 'react-native-paper';
+import styles from '../styles/Login';
 
 interface LoginProps {
   navigation: {
@@ -93,67 +93,3 @@ export default function Login({navigation}: LoginProps): React.JSX.Element {
   );
 }
 
-const styles = StyleSheet.create({
-  wrap: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: 'white',
-  },
-  input: {
-    height: 56,
-    width: 326,
-    borderColor: '#B5B5B5',
-    borderWidth: 1,
-    marginTop: 9,
-    paddingLeft: 21,
-  },
-  button: {
-    height: 56,
-    width: 326,
-    backgroundColor: '#D9D9D9',
-    borderRadius: 5,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  buttonText: {
-    color: '#939393',
-    fontSize: 18,
-    fontWeight: 'normal',
-  },
-  checkboxContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'flex-start',
-    width: 320,
-    marginTop: 14,
-  },
-  checkboxText: {
-    color: '#AAA',
-    fontSize: 14,
-    fontStyle: 'normal',
-    fontWeight: '400',
-  },
-  bottomTextLine: {
-    flexDirection: 'row',
-    width: 326,
-    justifyContent: 'center',
-    gap: 10,
-    marginTop: 26,
-  },
-  bottomText: {
-    color: '#7F7F7F',
-    textAlign: 'right',
-    fontSize: 14,
-    fontStyle: 'normal',
-    fontWeight: '400',
-  },
-  bottomTextSign: {
-    color: '#7F7F7F',
-    textAlign: 'right',
-    fontSize: 14,
-    fontStyle: 'normal',
-    fontWeight: '400',
-    marginTop: 1.4,
-  },
-});
