+++ frontend/src/pages/Store.tsx
@@ -1,6 +1,7 @@
 import React, { useState } from 'react';
-import { StyleSheet, View, Text, TouchableOpacity, ScrollView } from 'react-native';
+import {View, Text, TouchableOpacity, ScrollView } from 'react-native';
 import { NavigationProp } from '../navigation/NavigationProps'; // 인터페이스 import
+import styles from '../styles/Store';
 
 export default function Store({ navigation }: NavigationProp): React.JSX.Element {
   const [isFastOrderOn, setIsFastOrderOn] = useState(true);
@@ -57,113 +58,3 @@ export default function Store({ navigation }: NavigationProp): React.JSX.Element
   );
 }
 
-const styles = StyleSheet.create({
-  wrap: {
-    flexGrow: 1,
-    justifyContent: 'flex-start',
-    alignItems: 'center',
-    marginTop: '0%',
-    paddingBottom : '49%'
-  },
-  storeImg: {
-    height: 157,
-    width: '100%',
-    backgroundColor: '#D9D9D9',
-  },
-  storeInfo: {
-    height: '21%',
-    width: '100%',
-    marginTop: 23,
-    marginLeft: 26.5,
-  },
-  InfoText: {
-    flexDirection: 'row',
-  },
-  storeName: {
-    color: '#484747',
-    fontSize: 20,
-    fontStyle: 'normal',
-    fontWeight: '600',
-  },
-  storeUniqueMenu: {
-    color: '#7C7C7C',
-    fontSize: 16,
-    fontStyle: 'normal',
-    fontWeight: '500',
-    marginLeft: 9,
-    marginTop : '0.5%'
-  },
-  storeAddress: {
-    color: '#464646',
-    fontSize: 14,
-    fontStyle: 'normal',
-    marginTop: 17,
-  },
-  storeOpen: {
-    marginTop: 10,
-    color: '#464646',
-    fontSize: 14,
-    fontStyle: 'normal',
-  },
-  storeFastOrder: {
-    color: '#7C7C7C',
-    fontSize: 12,
-    fontStyle: 'normal',
-    fontWeight: '500',
-    marginTop: '2.6%',
-  },
-  storePhoneNumber: {
-    color: '#464646',
-    fontSize: 14,
-    fontStyle: 'normal',
-    fontWeight: '500',
-    marginTop: 7,
-  },
-  storeMap: {
-    width: 141,
-    height: 33,
-    borderRadius: 50,
-    backgroundColor: '#D9D9D9',
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  storeMapText: {
-    color: '#464646',
-    fontSize: 16,
-    fontStyle: 'normal',
-    fontWeight: '600',
-  },
-  MapWrap: {
-    width: '92%',
-    alignItems: 'center',
-    marginTop: 24,
-  },
-  padding: {
-    width: '100%',
-    height: 10,
-    backgroundColor: '#D9D9D9',
-  },
-  menu: {
-    width: '100%',
-    height: '12%',
-    borderWidth: 1,
-    borderTopWidth: 0,
-    borderColor: '#DFDFDF',
-  },
-  menuName: {
-    color : '#484747',
-    fontSize : 18,
-    fontStyle : 'normal',
-    fontWeight : '600',
-    paddingLeft: '6%',
-    paddingTop: '6%',
-  },
-  menuPrice: {
-    color : '#484747',
-    fontSize : 16,
-    fontStyle : 'normal',
-    fontWeight : '500',
-    paddingLeft: '6%',
-    marginTop: '5%',
-  },
-});
