+++ frontend/src/styles/SignUp.js
@@ -1,0 +1,12 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+    wrap : {
+        flex : 1,
+        color : 'black',
+        justifyContent : 'center',
+        alignItems : 'center'
+    }
+});
+
+export default styles;
